Implemented the text under the animation
Changed set_animation to accept a text_timeline, which is actually
an array of strings, and a text_id, which is the id of the element
where the text should be placed. To support text, frames in
timeline objects should have a text variable indicating the line
of text in text_timeline to use.

The WindowGroomer animation conforms to this format.